ml/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
google-generativeai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
cloudinary